e2e: node: cpumgr: replace old testsuite
This final change in the series completes the transition
to the new test suite
